Conflicts:
	Cares.Web/Cares.Web.csproj
	Cares.Web/ModelMappers/CompanyMapper.cs
	Cares.Web/Web.config
	Interfaces/Repository/IBusinessLegalStatusRepository.cs
	Repository/TypeRegistrations.cs

+++ Interfaces/Repository/IBpRatingTypeRepository.cs
@@ -5,7 +5,7 @@ namespace Interfaces.Repository
     /// <summary>
     /// Business Partner Rating Type Repository Interface
     /// </summary>
-    public interface IBpRatingTypeRepository : IBaseRepository<BpRatingType, int>
+    public interface IBpRatingTypeRepository : IBaseRepository<BpRatingType, long>
     {
     }
 }
