+++ Interfaces/Repository/IMeasurementUnit.cs
@@ -5,7 +5,7 @@ namespace Interfaces.Repository
     /// <summary>
     /// Measurement Unit Repository Interface
     /// </summary>
-    public interface IMeasurementUnit : IBaseRepository<MeasurementUnit, int>
+    public interface IMeasurementUnit : IBaseRepository<MeasurementUnit, long>
     {
 
     }
